修复图像标签模糊的问题 (#339)

+++ qfluentwidgets/components/widgets/progress_ring.py
@@ -119,21 +119,21 @@ class IndeterminateProgressRing(QProgressBar):
         if start:
             self.start()
 
-    @pyqtProperty(float)
+    @pyqtProperty(int)
     def startAngle(self):
         return self._startAngle
 
     @startAngle.setter
-    def startAngle(self, angle: float):
+    def startAngle(self, angle: int):
         self._startAngle = angle
         self.update()
 
-    @pyqtProperty(float)
+    @pyqtProperty(int)
     def spanAngle(self):
         return self._spanAngle
 
     @spanAngle.setter
-    def spanAngle(self, angle: float):
+    def spanAngle(self, angle: int):
         self._spanAngle = angle
         self.update()
